services/reconciliation/pyproject.toml
# reconciliation — bounded context. Exposes /reconciliation for the dashboard
# and a scheduled job that materializes the joined view for Looker (PRD §2.2).
[project]
name = "msm-reconciliation"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "google-cloud-bigquery>=3.25",
    "structlog>=24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "import-linter>=2.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
addopts = "--cov=msm_reconciliation --cov-fail-under=80"
testpaths = ["tests"]
